Refine home layout with shared ambient hero
